# bthwani-backend-nest/reports/openapi.yaml
openapi: 3.0.0
paths:
...
  /content/banners:
    get:
      operationId: ContentController_getActiveBanners
      parameters:
        - name: placement
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: الحصول على البانرات النشطة (public)
      tags:
        - Content
    post:
      operationId: ContentController_createBanner
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBannerDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: إنشاء بانر جديد
      tags:
        - Content
  /content/banners/{id}/click:
    post:
      operationId: ContentController_recordBannerClick
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: ''
      summary: تسجيل نقرة على بانر
      tags:
        - Content
  /content/admin/banners:
    get:
      operationId: ContentController_getAllBanners
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: كل البانرات (admin)
      tags:
        - Content
  /content/banners/{id}:
    patch:
      operationId: ContentController_updateBanner
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBannerDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تحديث بانر
      tags:
        - Content
    delete:
      operationId: ContentController_deleteBanner
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: حذف بانر
      tags:
        - Content
  /content/stores/{storeId}/sections:
    get:
      operationId: ContentController_getStoreSections
      parameters:
        - name: storeId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: الحصول على أقسام المتجر (public)
      tags:
        - Content
  /content/sections:
    post:
      operationId: ContentController_createStoreSection
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreSectionDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: إنشاء قسم متجر
      tags:
        - Content
  /content/sections/{id}:
    patch:
      operationId: ContentController_updateStoreSection
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreSectionDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تحديث قسم
      tags:
        - Content
    delete:
      operationId: ContentController_deleteStoreSection
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: حذف قسم
      tags:
        - Content
  /content/subscription-plans:
    get:
      operationId: ContentController_getSubscriptionPlans
      parameters: []
      responses:
        '200':
          description: ''
      summary: الحصول على خطط الاشتراك (public)
      tags:
        - Content
    post:
      operationId: ContentController_createSubscriptionPlan
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionPlanDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: إنشاء خطة اشتراك
      tags:
        - Content
  /content/subscribe:
    post:
      operationId: ContentController_subscribe
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscribeDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: الاشتراك في خطة
      tags:
        - Content
  /content/my-subscription:
    get:
      operationId: ContentController_getMySubscription
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: الحصول على اشتراكي
      tags:
        - Content
  /content/my-subscription/cancel:
    patch:
      operationId: ContentController_cancelSubscription
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: إلغاء الاشتراك
      tags:
        - Content
  /content/pages:
    get:
      operationId: ContentController_getCMSPages
      parameters:
        - name: type
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: الحصول على صفحات CMS (public)
      tags:
        - Content
  /content/pages/{slug}:
    get:
      operationId: ContentController_getCMSPageBySlug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: الحصول على صفحة CMS بالـ slug
      tags:
        - Content
  /content/admin/pages:
    post:
      operationId: ContentController_createCMSPage
      parameters: []
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: إنشاء صفحة CMS
      tags:
        - Content
  /content/admin/pages/{id}:
    patch:
      operationId: ContentController_updateCMSPage
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تحديث صفحة CMS
      tags:
        - Content
...
  /content/admin/app-settings:
    patch:
      operationId: ContentController_updateAppSettings
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تحديث إعدادات التطبيق
      tags:
        - Content
  /content/faqs:
    get:
      operationId: ContentController_getFAQs
      parameters:
        - name: category
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: الأسئلة الشائعة (public)
      tags:
        - Content
  /content/admin/faqs:
    post:
      operationId: ContentController_createFAQ
      parameters: []
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      summary: إضافة سؤال شائع
      tags:
        - Content
  /content/admin/faqs/{id}:
    patch:
      operationId: ContentController_updateFAQ
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: تحديث سؤال شائع
      tags:
        - Content
    delete:
      operationId: ContentController_deleteFAQ
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      summary: حذف سؤال شائع
      tags:
        - Content
...
info:
  title: Bthwani API v2
  description: NestJS API Documentation - نظام إدارة الطلبات والتجارة
  version: '2.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateBannerDto:
      type: object
      properties:
        title:
          type: string
          description: العنوان
          example: عرض خاص
        description:
          type: string
          description: الوصف
          example: خصم 50% على جميع المنتجات
        image:
          type: string
          description: الصورة
          example: https://example.com/banner.jpg
        link:
          type: string
          description: الرابط
          example: /products/123
        store:
          type: string
          description: معرف المتجر
        category:
          type: string
          description: معرف الفئة
        placement:
          type: string
          description: موضع العرض
          enum:
            - home
            - category
            - store
            - search
          example: home
        order:
          type: number
          description: الترتيب
          example: 1
          default: 0
        startDate:
          format: date-time
          type: string
          description: تاريخ البداية
          example: '2025-01-01'
        endDate:
          format: date-time
          type: string
          description: تاريخ النهاية
          example: '2025-12-31'
      required:
        - image
        - order
    UpdateBannerDto:
      type: object
      properties: {}
    CreateStoreSectionDto:
      type: object
      properties:
        store:
          type: string
          description: معرف المتجر
        name:
          type: string
          description: الاسم
          example: خضروات
        nameAr:
          type: string
          description: الاسم بالعربية
          example: خضروات
        description:
          type: string
          description: الوصف
          example: قسم الخضروات الطازجة
        icon:
          type: string
          description: الأيقونة
          example: 🥕
        usageType:
          type: string
          description: نوع الاستخدام
          enum:
            - grocery
            - restaurant
            - retail
          example: grocery
        order:
          type: number
          description: الترتيب
          example: 1
          default: 0
      required:
        - store
        - name
        - usageType
        - order
    UpdateStoreSectionDto:
      type: object
      properties: {}
    CreateSubscriptionPlanDto:
      type: object
      properties: {}
    SubscribeDto:
      type: object
      properties: {}
...